+++ 03-api/src/main/scala/camundala/api/ApiBaseDsl.scala
@@ -8,7 +8,7 @@ import scala.reflect.ClassTag
 
 trait ApiBaseDsl:
 
-  def group(name: String)(apis: CApi*): CApiGroup =
+  def group(name: String)(apis: InOutApi[?,?]*): CApiGroup =
     CApiGroup(name, apis.toList)
 
   def api[
@@ -21,7 +21,7 @@ trait ApiBaseDsl:
   def api[
       In <: Product: InOutEncoder: InOutDecoder: Schema,
       Out <: Product: InOutEncoder: InOutDecoder: Schema: ClassTag
-  ](pApi: ProcessApi[In, Out])(body: CApi*): ProcessApi[In, Out] =
+  ](pApi: ProcessApi[In, Out])(body: InOutApi[?,?]*): ProcessApi[In, Out] =
     pApi.withApis(body.toList)
 
   extension [