+++ 03-worker/src/main/scala/camundala/worker/EngineContext.scala
@@ -2,6 +2,7 @@ package camundala
 package worker
 
 import camundala.domain.*
+import camundala.bpmn.*
 import camundala.worker.CamundalaWorkerError.ServiceError
 
 import java.time.LocalDateTime